requirements.txt
pandas
torch
torchvision
transformers
matplotlib
Pillow
scikit-learn 

torch==2.9.0
torchvision==0.24.0
fastapi==0.68.0
uvicorn==0.15.0
pillow==8.4.0
streamlit==1.4.0
